requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
beautifulsoup4==4.10.0
branca==0.4.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-rest-auth==2.1.11
Django==3.2.7
django-bootstrap4==3.0.1
django-bulma==0.8.2
django-cors-headers==3.9.0
django-countries==7.2.1
django-filter==21.1
django-import-export==2.6.1
django-jazzmin==2.4.8
django-leaflet==0.28.1
django-pandas==0.6.4
django-templated-mail==1.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
et-xmlfile==1.1.0
folium==0.12.1
future==0.18.2
html5lib==1.1
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupPy==1.14
MarkupSafe==2.0.1
numpy==1.21.2
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.3.3
Pillow==8.3.2
pycparser==2.20
PyJWT==2.1.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
reportlab==3.6.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sorl-thumbnail==12.7.0
soupsieve==2.2.1
sqlparse==0.4.2
tablib==3.0.0
uritemplate==3.0.1
urllib3==1.26.7
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
